routine

+++ dotfiles/fish/functions/poetry_new.fish
@@ -1,12 +1,16 @@
 function poetry_new
-    poetry new $argv
-    cd $argv
-    rm -rf tests
-    rm -rf $argv
-    rm README.rst
-    echo /.venv/ > .gitignore
-    touch README.md
-    poetry install
-    git init
-    cd ..
+    if test (count $argv) -ne 0
+        poetry new $argv
+        cd $argv
+        rm -rf tests
+        rm -rf $argv
+        rm README.rst
+        echo /.venv/ > .gitignore
+        touch README.md
+        poetry install
+        git init
+        cd ..
+    else
+        echo 'argv error'
+    end
 end
